requirements.txt
gradio
phidata==2.7.3
Pillow==10.0.0
duckduckgo-search
google-generativeai==0.8.3